Add LLM advisory scheduling to xv6 kernel
Introduces a mechanism for external agents to advise the kernel scheduler via a new set_llm_advice syscall. Adds llmhelper user program to inject advice, updates process statistics, and logs scheduling state for agent consumption. Includes Makefile, syscall, and userland integration, as well as runner.py for host automation.

+++ xv6/kernel/sysproc.c
@@ -1,3 +1,4 @@
+// kernel/sysproc.c
 #include "types.h"
 #include "riscv.h"
 #include "defs.h"
@@ -49,16 +50,14 @@ sys_sbrk(void)
 
   if(t == SBRK_EAGER || n < 0) {
     if(growproc(n) < 0) {
-      return -1;
+      return (uint64)-1;
     }
   } else {
-    // Lazily allocate memory for this process: increase its memory
-    // size but don't allocate memory. If the processes uses the
-    // memory, vmfault() will allocate it.
+    // Lazy growth.
     if(addr + n < addr)
-      return -1;
+      return (uint64)-1;
     if(addr + n > TRAPFRAME)
-      return -1;
+      return (uint64)-1;
     myproc()->sz += n;
   }
   return addr;
@@ -73,12 +72,13 @@ sys_pause(void)
   argint(0, &n);
   if(n < 0)
     n = 0;
+
   acquire(&tickslock);
   ticks0 = ticks;
-  while(ticks - ticks0 < n){
+  while(ticks - ticks0 < (uint)n){
     if(killed(myproc())){
       release(&tickslock);
-      return -1;
+      return (uint64)-1;
     }
     sleep(&ticks, &tickslock);
   }
@@ -90,18 +90,15 @@ uint64
 sys_kill(void)
 {
   int pid;
-
   argint(0, &pid);
   return kkill(pid);
 }
 
-// return how many clock tick interrupts have occurred
-// since start.
+// return how many clock tick interrupts have occurred since start.
 uint64
 sys_uptime(void)
 {
   uint xticks;
-
   acquire(&tickslock);
   xticks = ticks;
   release(&tickslock);
